requirements.txt
Django==4.2.19
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-heroku==0.3.1
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.10
drf-yasg==1.21.10
PyJWT==2.9.0
Pillow==7.2.0
pytz
